Cambios conexion base de datos y creacion de dominio

+++ application/commands/user_comand.go
@@ -1,7 +1,7 @@
 package commands
 
 type UserCommand struct {
-	DNI            int16  `json:"dni"`
+	DNI            int64  `json:"dni"`
 	Name           string `json:"name"`
 	LastName       string `json:"last_name"`
 	QuantityMovies int8   `json:"quantity_movies"`
